Fix TableHead indent
